Change the quarter correction logic to be a bit smarter.
Before it was just using the quarter of the last play in the drive
according to the JSON. But this isn't always correct. Instead, use
the maximum quarter number of all the plays in the drive. If that
doesn't work out, use a heuristic to increase the quarter number by
one when the drive start time is greater than the drive end time
(and only in quarters 1 and 3, of course).